Separated Range and Slice into own headers.
